Track the year of a file creation through license-maven-plugin-git
Reason behind this PR:
Sometimes you are required to include in your license header only the
date of the file creation.
These changes allow us to look at the git history and find the year of
reation for a given file.

Note:
Sometimes git is unable to figure out the date of creation of a file if
this has moved/renamed without git being able to detect such change.
A potential future solution could be to avoid dates to be overwritten.
I would leave that for another PR though.

Additionally, this PR reduces a bit of duplicated code in the git plugin.