Support injecting Copyright Years based on file existence
For some Copyright policies the years in the header should reflect only
the existence of that file not the project as a whole.  While this can
be achieved by using ${license.git.copyrightCreationYear} and
${license.git.copyrightLastYear} in the header template you end up
generating awkward headers for some files e.g. 2021-2021.  And you
cannot use ${license.git.copyrightYears} for this use case because that
will use the project inception year giving you something like 2019-2021.

This commit adds a new calculated property
${license.git.copyrightExistenceYears} which provides this information
and may be used in header templates where this behaviour is desired.

It also fixes a bug in GitLookup whereby if you had modified the file
it would incorrectly calculate the year range (because the creation year
was always defaulted to current year for modified files) and falsely flag
a license header as missing (or reformat unncessarily).