[ci] Drop travis CI
rate limited, no longer open source friendly, and github actions is light years ahead.

Basic github actions added that does all that is necessary.  I'm not 100% clear why we are testing the release deployment features here but all that can be done via actions if we need something similar and probably only a few minutes worth of work.  Dropping it as travis is just bad at this point.