fix: Configure the project for the `ProjectBuildingRequest` being used.
The `ProjectBuildingRequest` requires a project to be configured. This project is not available by default. The `ProjectBuildingRequest` is mutable, so I did not use the exiting class variable. Instead, I suggest not storing the request but create one as needed. For that purpose, the Maven session is now kept around.

When requiring a request with a project, the existing copy-constructor is new used and a project is configured as needed.

This change also fixes the tests (expecting 0 instead of 1 dependencies). This change has little effect though, as the `MavenProjectLicensesIT` does not actually run.

This fixes #377.