refactor: Use Maven Integration Test framework to run some integration tests.
This makes the `MavenProjectLicensesIT` actually run during verification. The testing framework by Karl Heinz Marbaise takes care of resource copying and filtering and also manual verifcation can be simplified.

The biggest advantage of this approach is not having to locally install snapshot variants all the time and a streamlined process.

Debugging is still possible (see link in the JavaDoc) via remote debugger.