[headerType] Newline on the end for slashstar_style
If you see examples in java, kotlin, etc., they mostly have newline after the copyright statement.
https://github.com/mathieucarbou/license-maven-plugin/pull/186
It might be because Intellij applies header file in that way and also because in this way this would be easier to divide copyright and the code.
Since customizing the headerDefinition regarding ending newline makes a quite long configuration having to reconfigure all other values, making it default might be better ƒor this plugin to be used more commonly.