[cleanup] Remove internally deprecated default class not used
All this does is waste memory on the license-maven-plugin-fs test that uses it.

The adjustProperties fails to use what it created there so that just goes to GC.

The deprecated method only returns Collections.emptyMap().

Due to test in the license-maven-plugin-fs, keep these as default but have them do nothing as that is what they are actually doing.  The adjustProperties will remain to send back emptyMap()